# docs/swagger.yaml
basePath: /
definitions:
  domain.Test:
    properties:
      author_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title:
        maxLength: 255
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        maxLength: 255
        minLength: 3
        type: string
      id:
        type: integer
      name:
        maxLength: 255
        minLength: 3
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    - password
    type: object
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
  v1.signInResponse:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Qna API
  title: Qna API
  version: "1.0"
paths:
  /sign-in:
    post:
      consumes:
      - application/json
      description: Sign in
      operationId: sign-in
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.signInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign in
      tags:
      - auth
  /sign-up:
    post:
      consumes:
      - application/json
      description: Sign up
      operationId: sign-up
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign up
      tags:
      - auth
  /tests:
    get:
      consumes:
      - application/json
      description: Get all tests by current user
      operationId: get-all-tests-by-current-user
      parameters:
      - description: page id
        in: query
        name: page_id
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Test'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all tests by current user
      tags:
      - tests
    post:
      consumes:
      - application/json
      description: Create test
      operationId: create-test
      parameters:
      - description: test
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/domain.Test'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create test
      tags:
      - tests
  /tests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete test by id
      operationId: delete-test-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete test by id
      tags:
      - tests
    get:
      consumes:
      - application/json
      description: Get test by id
      operationId: get-test-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Test'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get test by id
      tags:
      - tests
    put:
      consumes:
      - application/json
      description: Update test by id
      operationId: update-test-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: test
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/domain.Test'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update test by id
      tags:
      - tests
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"